Refactor invite handling and update routes
- Update `InvitesController` to handle invites without project context
- Modify `accept` method to use `@invite.project`
- Change `set_invite` to find invite directly
- Adjust `ensure_can_respond!` to check email instead of user
- Move invite modal logic to new `Invite` component
- Remove console log from `Projects.jsx`
- Add destroy route for invites in `routes.rb`
